model_service/requirements.txt
flask==3.0.0
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.26.0
gunicorn==21.2.0
